refactor: rename addClassNameToComponent to withClassName
